# perf: use `Cow` in `EitherString`