Seems like a small win in `dict_json` benchmark.

Before

```
test dict_json         ... bench:      14,610 ns/iter (+/- 215)
test dict_key_error    ... bench:      11,396 ns/iter (+/- 147)
test dict_python       ... bench:       5,780 ns/iter (+/- 114)
test ints_json         ... bench:          26 ns/iter (+/- 1)
test ints_python       ... bench:          12 ns/iter (+/- 0)
test list_any_json     ... bench:       2,434 ns/iter (+/- 19)
test list_any_python   ... bench:         907 ns/iter (+/- 14)
test list_error_json   ... bench:       3,121 ns/iter (+/- 17)
test list_error_python ... bench:       3,595 ns/iter (+/- 22)
test list_int_json     ... bench:       2,464 ns/iter (+/- 14)
test list_int_python   ... bench:       1,892 ns/iter (+/- 28)
test model_json        ... bench:       1,589 ns/iter (+/- 9,252)
test model_python      ... bench:         898 ns/iter (+/- 15,014)
```

After:

```
test dict_json         ... bench:      13,721 ns/iter (+/- 207)
test dict_key_error    ... bench:      11,459 ns/iter (+/- 70)
test dict_python       ... bench:       5,457 ns/iter (+/- 193)
test ints_json         ... bench:          26 ns/iter (+/- 1)
test ints_python       ... bench:          12 ns/iter (+/- 0)
test list_any_json     ... bench:       2,484 ns/iter (+/- 11)
test list_any_python   ... bench:         922 ns/iter (+/- 10)
test list_error_json   ... bench:       3,127 ns/iter (+/- 296)
test list_error_python ... bench:       3,650 ns/iter (+/- 47)
test list_int_json     ... bench:       2,476 ns/iter (+/- 12)
test list_int_python   ... bench:       1,961 ns/iter (+/- 51)
test model_json        ... bench:       1,543 ns/iter (+/- 64)
test model_python      ... bench:         891 ns/iter (+/- 9,924)
```